% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_ducklake.R
\name{attach_ducklake}
\alias{attach_ducklake}
\title{Attach, create or update a ducklake}
\usage{
attach_ducklake(
  conduckdb,
  ducklake_filename = NULL,
  ducklake_alias = NULL,
  create_if_not_exists = FALSE,
  read_only = TRUE,
  ducklake_type = "duckdb",
  ducklake_encrypt = FALSE,
  parquet_encrypt = FALSE,
  parquet_directory = "data_files",
  override_parquet_directory = FALSE
)
}
\arguments{
\item{conduckdb}{an existing duckdb connection}

\item{ducklake_filename}{file name (and directory if not in working directory) of ducklake}

\item{ducklake_alias}{short alias to use instead of ducklake_filename when referring to ducklake database}

\item{create_if_not_exists}{create a new ducklake if it does not already exist (default is FALSE)}

\item{read_only}{connect read only to the ducklake. if create_if_not_exists is TRUE this should be FALSE (default is TRUE)}

\item{ducklake_type}{which SQL software to use as catalog for metadata (duckdb or sqlite)}

\item{ducklake_encrypt}{should the ducklake itself be encrypted (for duckdb only - default is TRUE)}

\item{parquet_encrypt}{should the parquet files storing tables be encrypted (default is TRUE)}

\item{parquet_directory}{directory where the parquet files are / should be stored (default is "data_files")}

\item{override_parquet_directory}{if data path is different to default should this overrides the default (default is FALSE)}
}
\value{
Attaches the ducklake to your duckdb connection and with message confirming the ducklake is the default database.
Returns dataframe of attached databases.
}
\description{
Attach, create or update a ducklake
}
\examples{
my_con <- connect_duckdb()

attach_ducklake(my_con,
   ducklake_filename = tempfile(pattern = "metadata.ducklake"),
   ducklake_alias = "my_ducklake",
   create_if_not_exists = TRUE,
   ducklake_encrypt = FALSE,
   parquet_directory = tempdir())

switch_default_database(my_con, "memory")

switch_default_database(my_con, "my_ducklake")

detach_ducklake(my_con, "my_ducklake", "memory")
}
