% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_ducklake.R
\name{attached_databases}
\alias{attached_databases}
\title{What databases are attached to the duckdb connection}
\usage{
attached_databases(conduckdb)
}
\arguments{
\item{conduckdb}{an existing duckdb connection}
}
\value{
a dataframe with the attached database names
}
\description{
What databases are attached to the duckdb connection
}
