% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_duckdb.R
\name{check_duckdb_connection}
\alias{check_duckdb_connection}
\title{Check the connection is a duckdb connection}
\usage{
check_duckdb_connection(conduckdb)
}
\arguments{
\item{conduckdb}{an existing duckdb connection}
}
\value{
NULL, no error means it is a dudkdb connection
}
\description{
Check the connection is a duckdb connection
}
\examples{
my_con <- connect_duckdb()
check_duckdb_connection(my_con)
disconnect_duckdb(my_con)
}
