% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_message.R
\name{commit_message}
\alias{commit_message}
\title{Add a message to the snapshot when updating a ducklake table etc.}
\usage{
commit_message(
  conduckdb,
  ducklake_alias,
  author,
  commit_message,
  commit_extra_info = NULL,
  code
)
}
\arguments{
\item{conduckdb}{an existing duckdb connection}

\item{ducklake_alias}{of the ducklake attached to the duckdb connection (character)}

\item{author}{author of the update}

\item{commit_message}{a message describing the update}

\item{commit_extra_info}{any extra information (default is NULL)}

\item{code}{the code making the change to the ducklake}
}
\value{
a dataframe containing snapshots to date
}
\description{
Add a message to the snapshot when updating a ducklake table etc.
}
