% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_duckdb.R
\name{connect_duckdb}
\alias{connect_duckdb}
\title{Create (or attach to an existing) duckdb database. This can be used to attach to the ducklake}
\usage{
connect_duckdb(duckdbname = NULL)
}
\arguments{
\item{duckdbname}{duckdb database name (default is memory)}
}
\value{
A duckdb database connection with two core extensions (ducklake and sqlite) installed and loaded
}
\description{
Create (or attach to an existing) duckdb database. This can be used to attach to the ducklake
}
\examples{
my_con <- connect_duckdb()
check_duckdb_connection(my_con)
disconnect_duckdb(my_con)
}
