% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_ducklake.R
\name{default_database}
\alias{default_database}
\title{Check the current default database on a duckdb connection}
\usage{
default_database(conduckdb)
}
\arguments{
\item{conduckdb}{an existing duckdb connection}
}
\value{
the name of the default database. Use \code{switch_default_database()} to change the default database
}
\description{
Check the current default database on a duckdb connection
}
