% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_duckdb.R
\name{disconnect_duckdb}
\alias{disconnect_duckdb}
\title{Disconnect from a duckdb connection}
\usage{
disconnect_duckdb(conduckdb)
}
\arguments{
\item{conduckdb}{an existing duckdb connection}
}
\value{
Disconnects from a duckdb connection
}
\description{
Disconnect from a duckdb connection
}
\examples{
my_con <- connect_duckdb()
check_duckdb_connection(my_con)
disconnect_duckdb(my_con)
}
