% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducklake_snapshots.R
\name{ducklake_snapshots}
\alias{ducklake_snapshots}
\title{Get details of updates (snapshots) to the ducklake}
\usage{
ducklake_snapshots(conduckdb, ducklake_alias)
}
\arguments{
\item{conduckdb}{an existing duckdb connection}

\item{ducklake_alias}{of the ducklake attached to the duckdb connection (character)}
}
\value{
a dataframe
}
\description{
Get details of updates (snapshots) to the ducklake
}
