% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducklake_table_info.R
\name{ducklake_table_info}
\alias{ducklake_table_info}
\title{Obtain information about the tables in the attached ducklake}
\usage{
ducklake_table_info(conduckdb, ducklake_alias, table_name_only = TRUE)
}
\arguments{
\item{conduckdb}{an existing duckdb connection}

\item{ducklake_alias}{of the ducklake attached to the duckdb connection (character)}

\item{table_name_only}{should the table names only or more detail be provided (default if TRUE)}
}
\value{
a dataframe
}
\description{
Obtain information about the tables in the attached ducklake
}
