% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_ducklake.R
\name{switch_default_database}
\alias{switch_default_database}
\title{Switch the default database on a duckdb connection}
\usage{
switch_default_database(conduckdb, new_default)
}
\arguments{
\item{conduckdb}{an existing duckdb connection}

\item{new_default}{the new_default database to switch to}
}
\value{
the new default database
}
\description{
Switch the default database on a duckdb connection
}
\examples{
my_con <- connect_duckdb()

attach_ducklake(my_con,
   ducklake_filename = tempfile(pattern = "metadata.ducklake"),
   ducklake_alias = "my_ducklake",
   create_if_not_exists = TRUE,
   ducklake_encrypt = FALSE,
   parquet_directory = tempdir())

switch_default_database(my_con, "memory")

switch_default_database(my_con, "my_ducklake")

detach_ducklake(my_con, "my_ducklake", "memory")
}
