% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_ducklake.R
\name{detach_ducklake}
\alias{detach_ducklake}
\title{Detach a ducklake from a duckdb connection}
\usage{
detach_ducklake(conduckdb, ducklake_alias, new_default = "memory")
}
\arguments{
\item{conduckdb}{an existing duckdb connection}

\item{ducklake_alias}{short alias to use instead of ducklake_filename when referring to ducklake database}

\item{new_default}{A new default database (default is memory).
Run \code{attached_databases()} for a list of currently attached databases}
}
\value{
detaches and returns remaining attached databases
}
\description{
Detach a ducklake from a duckdb connection
}
\examples{
my_con <- connect_duckdb()

attach_ducklake(my_con,
   ducklake_filename = tempfile(pattern = "metadata.ducklake"),
   ducklake_alias = "my_ducklake",
   create_if_not_exists = TRUE,
   read_only = FALSE,
   ducklake_encrypt = FALSE,
   parquet_directory = tempdir())

switch_default_database(my_con, "memory")

switch_default_database(my_con, "my_ducklake")

detach_ducklake(my_con, "my_ducklake", "memory")
}
